feat(build): add support for angular2 AOT compile
* add scripts for AOT build, cleaning, publishing
* fix AOT-related errors
* move build results to dedicated directories
* remove previous build results from source control
* cleanup published package
* upgrade angular to 2.1.0
* keep typescript to 2.0.3
* upgrade rxjs to 5.0.2
* add settings for VS Code
