--- .npmignore
@@ -1,4 +1,0 @@
-node_modules
-typings
-**/*.ts
-!index.d.ts
